+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add a dropdown on the "Receiver address" input in the send screen to select an account
 - Add feedback link to guide and about settings
 - Move active currencies to top of currency dropdown
+- Export logs now includes rotated log file (log.txt.1) when present
 - Android: fix connectivity misdetection when switching between WIFI and cellular network.
 - Android: dropped support for Android versions lower than 7.
 - Sort backups from newest to oldest in manage backups